+++ src/modules/variable/init.rs
@@ -1,16 +1,23 @@
 use heraclitus_compiler::prelude::*;
 use crate::context;
-use crate::modules::{Typed};
+use crate::modules::{Typed, Type};
 use crate::modules::expression::expr::Expr;
 use crate::translate::module::TranslateModule;
 use crate::utils::error::get_error_logger;
 use crate::utils::metadata::{ParserMetadata, TranslateMetadata};
-use super::{variable_name_extensions, handle_add_variable};
+use super::{variable_name_extensions, handle_identifier_name};
 
 #[derive(Debug, Clone)]
 pub struct VariableInit {
     name: String,
     expr: Box<Expr>
+}
+
+impl VariableInit {
+    fn handle_add_variable(&self, meta: &mut ParserMetadata, name: &str, kind: Type, tok: Option<Token>) {
+        handle_identifier_name(meta, name, tok);
+        meta.mem.add_variable(name, kind);
+    }
 }
 
 impl SyntaxModule<ParserMetadata> for VariableInit {
@@ -32,7 +39,7 @@ impl SyntaxModule<ParserMetadata> for VariableInit {
             token(meta, "=")?;
             syntax(meta, &mut *self.expr)?;
             // Add a variable to the memory
-            handle_add_variable(meta, &self.name, self.expr.get_type(), tok);
+            self.handle_add_variable(meta, &self.name, self.expr.get_type(), tok);
             Ok(())
         }, |details| {
             let message = format!("Expected '=' after variable name '{}'", self.name);
